pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zbxctl"
version = "0.3.0"
description = "Zabbix configuration as code — manage monitoring via Git"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["zabbix", "monitoring", "infrastructure-as-code", "devops", "gitops"]
authors = [{ name = "Pau Santana", email = "pausantanapi2@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.12",
    "requests>=2.31",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "paramiko>=3.4",
]

[project.urls]
Homepage = "https://github.com/psantana5/zbx"
Repository = "https://github.com/psantana5/zbx"
"Bug Tracker" = "https://github.com/psantana5/zbx/issues"
Changelog = "https://github.com/psantana5/zbx/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "responses>=0.25",
    "mypy>=1.8",
    "ruff>=0.3",
    "build>=1.0",
    "twine>=5.0",
]

[project.scripts]
zbx = "zbx.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["zbx*"]

[tool.setuptools.package-data]
zbx = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
